requirements.txt
fastapi==0.128.6
uvicorn==0.40.0
sqlmodel==0.0.32
httpx==0.28.1
jinja2==3.1.6
python-multipart==0.0.22
python-dotenv==1.2.1
itsdangerous==2.2.0
bcrypt==5.0.0
email-validator==2.3.0
Authlib==1.6.7
pydantic-settings==2.12.0
stripe==14.3.0
